+++ seadas-reader/src/main/java/gov/nasa/gsfc/seadas/dataio/L1AOctsFileReader.java
@@ -71,15 +71,15 @@ public class L1AOctsFileReader extends SeadasFileReader {
 
         int[] dims = lats.getShape();
 
-        double[] latTiePoints;
-        double[] lonTiePoints;
+        float[] latTiePoints;
+        float[] lonTiePoints;
 
         try {
             Array latarr = lats.read();
             Array lonarr = lons.read();
 
-            latTiePoints = (double[]) latarr.getStorage();
-            lonTiePoints = (double[]) lonarr.getStorage();
+            latTiePoints = (float[]) latarr.getStorage();
+            lonTiePoints = (float[]) lonarr.getStorage();
 
             final TiePointGrid latGrid = new TiePointGrid("latitude", dims[1], dims[0], 0, 0,
                     susampX, subsampY, latTiePoints);
