+++ seadas-reader/src/main/java/gov/nasa/gsfc/seadas/dataio/L1BModisFileReader.java
@@ -134,7 +134,7 @@ public class L1BModisFileReader extends SeadasFileReader {
                         float_data_raw = (short[]) dataArray.getStorage();
                     }
 
-                    final double[] float_data = new double[(int) dataArray.getSize()];
+                    final float[] float_data = new float[(int) dataArray.getSize()];
                     for (int i = 0; i < float_data.length; i++) {
                         float_data[i] = float_data_raw[i] * scale_factor;
                     }
@@ -315,16 +315,16 @@ public class L1BModisFileReader extends SeadasFileReader {
 
             //Use lat/lon with TiePointGeoCoding
             int[] dims = lats.getShape();
-            double[] latTiePoints;
-            double[] lonTiePoints;
+            float[] latTiePoints;
+            float[] lonTiePoints;
             Array latarr = lats.read();
             Array lonarr = lons.read();
             if (mustFlipX && mustFlipY) {
-                latTiePoints = (double[]) latarr.flip(0).flip(1).copyTo1DJavaArray();
-                lonTiePoints = (double[]) lonarr.flip(0).flip(1).copyTo1DJavaArray();
+                latTiePoints = (float[]) latarr.flip(0).flip(1).copyTo1DJavaArray();
+                lonTiePoints = (float[]) lonarr.flip(0).flip(1).copyTo1DJavaArray();
             } else {
-                latTiePoints = (double[]) latarr.getStorage();
-                lonTiePoints = (double[]) lonarr.getStorage();
+                latTiePoints = (float[]) latarr.getStorage();
+                lonTiePoints = (float[]) lonarr.getStorage();
             }
 
             if (externalGeo) {
